+++ app/renderer/settings.js
@@ -1,6 +1,6 @@
 /*
 * Vieb - Vim Inspired Electron Browser
-* Copyright (C) 2019-2025 Jelmer van Arnhem
+* Copyright (C) 2019-2026 Jelmer van Arnhem
 *
 * This program is free software: you can redistribute it and/or modify
 * it under the terms of the GNU General Public License as published by
@@ -55,6 +55,7 @@ const mouseFeatures = [
     "pageoutsideinsert",
     "switchtab",
     "history",
+    "historyswipe",
     "guiontop",
     "newtab",
     "closetab",